+++ src/Symfony/Component/DependencyInjection/LazyProxy/ProxyHelper.php
@@ -19,54 +19,6 @@ namespace Symfony\Component\DependencyInjection\LazyProxy;
 class ProxyHelper
 {
     /**
-     * @return string The signature of the passed function, return type and function/method name included if any
-     */
-    public static function getSignature(\ReflectionFunctionAbstract $r, &$call = null)
-    {
-        $signature = array();
-        $call = array();
-
-        foreach ($r->getParameters() as $i => $p) {
-            $k = '$'.$p->name;
-            if (method_exists($p, 'isVariadic') && $p->isVariadic()) {
-                $k = '...'.$k;
-            }
-            $call[] = $k;
-
-            if ($p->isPassedByReference()) {
-                $k = '&'.$k;
-            }
-            if ($type = self::getTypeHint($r, $p)) {
-                $k = $type.' '.$k;
-            }
-            if ($type && $p->allowsNull()) {
-                $k = '?'.$k;
-            }
-
-            try {
-                $k .= ' = '.self::export($p->getDefaultValue());
-                if ($type && $p->allowsNull() && null === $p->getDefaultValue()) {
-                    $k = substr($k, 1);
-                }
-            } catch (\ReflectionException $e) {
-                if ($type && $p->allowsNull() && !class_exists('ReflectionNamedType', false)) {
-                    $k .= ' = null';
-                    $k = substr($k, 1);
-                }
-            }
-
-            $signature[] = $k;
-        }
-        $call = ($r->isClosure() ? '' : $r->name).'('.implode(', ', $call).')';
-
-        if ($type = self::getTypeHint($r)) {
-            $type = ': '.($r->getReturnType()->allowsNull() ? '?' : '').$type;
-        }
-
-        return ($r->returnsReference() ? '&' : '').($r->isClosure() ? '' : $r->name).'('.implode(', ', $signature).')'.$type;
-    }
-
-    /**
      * @return string|null The FQCN or builtin name of the type hint, or null when the type hint references an invalid self|parent context
      */
     public static function getTypeHint(\ReflectionFunctionAbstract $r, \ReflectionParameter $p = null, $noBuiltin = false)