+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services9.php
@@ -34,7 +34,6 @@ class ProjectServiceContainer extends Container
         $this->methodMap = array(
             'bar' => 'getBarService',
             'baz' => 'getBazService',
-            'closure_proxy' => 'getClosureProxyService',
             'configurator_service' => 'getConfiguratorServiceService',
             'configurator_service_simple' => 'getConfiguratorServiceSimpleService',
             'configured_service' => 'getConfiguredServiceService',
@@ -107,21 +106,6 @@ class ProjectServiceContainer extends Container
         $instance->setFoo(${($_ = isset($this->services['foo_with_inline']) ? $this->services['foo_with_inline'] : $this->get('foo_with_inline')) && false ?: '_'});
 
         return $instance;
-    }
-
-    /**
-     * Gets the 'closure_proxy' service.
-     *
-     * This service is shared.
-     * This method always returns the same instance of the service.
-     *
-     * @return \BarClass A BarClass instance
-     */
-    protected function getClosureProxyService()
-    {
-        return $this->services['closure_proxy'] = new \BarClass(/** @closure-proxy BarClass::getBaz */ function () {
-            return ${($_ = isset($this->services['closure_proxy']) ? $this->services['closure_proxy'] : $this->get('closure_proxy')) && false ?: '_'}->getBaz();
-        });
     }
 
     /**